+++ src/app/account/account.module.ts
@@ -22,9 +22,11 @@ import { PaginationModule } from "../common/components/pagination-component/pagi
 import { FlowsViewModule } from "../dataset-flow/flows-view.module";
 import { FeatureFlagModule } from "../common/directives/feature-flag.module";
 import { MatProgressBarModule } from "@angular/material/progress-bar";
+import { SettingsTabComponent } from "./additional-components/settings-tab/settings-tab.component";
+import { AccountSettingsModule } from "./settings/account-settings.module";
 
 @NgModule({
-    declarations: [AccountComponent, AccountFlowsTabComponent, DatasetsTabComponent],
+    declarations: [AccountComponent, AccountFlowsTabComponent, DatasetsTabComponent, SettingsTabComponent],
     imports: [
         CommonModule,
         FormsModule,
@@ -37,6 +39,7 @@ import { MatProgressBarModule } from "@angular/material/progress-bar";
         ReactiveFormsModule,
         RouterModule,
 
+        AccountSettingsModule,
         DatasetListModule,
         FeatureFlagModule,
         FlowsViewModule,